quick reorganizing
